fix: moved upgradeContract from task utils to test utils to avoid importing hardhat before its config is loaded
